Move styles into their own files
