+++ client/src/components/DeactivateAccountButton.js
@@ -3,8 +3,14 @@
  */
 
 import React, { useEffect,useState } from 'react'
+import { styled } from '@material-ui/styles';
 import { Button } from '@material-ui/core';
 import axios from 'axios';
+
+/**
+ * Define styles
+ */
+
 
 /**
  * Define component
