+++ client/src/views/styles/HeaderH2.js
@@ -1,0 +1,21 @@
+/**
+ * Dependencies
+ */
+
+import { styled } from '@material-ui/styles';
+
+/**
+ * Define style
+ */
+
+const HeaderH2 = styled('h2')({
+  textAlign: 'center',
+  padding: '10px 0px',
+  color: 'white',
+});
+
+/**
+ * Export style
+ */
+
+export default HeaderH2;
