+++ client/src/views/styles/index.js
@@ -1,0 +1,9 @@
+/**
+ * Export styles
+ */
+
+export { default as HeaderH1 } from './HeaderH1';
+export { default as HeaderH2 } from './HeaderH2';
+export { default as HeaderH3 } from './HeaderH3';
+export { default as BecomeMediatorLink } from './BecomeMediatorLink';
+export { default as EditUserLink } from './EditUserLink';
